2/2 clean and edit
Formatting tabs from source to markdown, the only consistent result I am getting to achieve the same effect is to apply 4 spaces to replace a tab. This simply takes too much effort to apply to EVERY SINGLE solution. Readability it slightly off, but it is still ok. Many of the query step explanations had to be edited to be bolded to differentiate from the sea of regular text.